FlightAgent/orchestrator/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
langchain==0.3.7
langchain-openai==0.2.7
pydantic==2.8.2
requests==2.31.0
redis==5.0.5
sse-starlette==1.6.1
opentelemetry-sdk==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-exporter-otlp-proto-http==1.30.0
